// src/dtos/adminOwner/carOwnerVerifyList.response.dto.ts
// owner-verify-list.response.dto.ts
export interface OwnerVerifyListItemDTO {
  id: string;
  fullName: string;
  email: string;
  phoneNumber?: string;
  verifyStatus: number;
  blockStatus: number;
  idVerified: boolean;
  rejectionReason?: string;
  createdAt?: Date;
}

export interface OwnerVerifyListResponseDTO {
  carOwners: OwnerVerifyListItemDTO[];
  total: number;
}